Chore/unit test

+++ backend/rpidrive/controllers/user.py
@@ -4,10 +4,14 @@ from django.contrib.auth.models import User
 from django.db import transaction
 from django.db.models import QuerySet
 
-from rpidrive.controllers.exceptions import NoPermissionException
+from rpidrive.controllers.exceptions import (
+    InvalidOperationRequestException,
+    NoPermissionException,
+    ObjectNotFoundException,
+)
 
 
-class UserNotFoundException(Exception):
+class UserNotFoundException(ObjectNotFoundException):
     """User not found exception"""
 
 
@@ -30,7 +34,7 @@ def create_user(
         raise InvalidUsernameException("Invalid username.")
     password = password.strip()
     if not password:
-        raise InvalidUsernameException("Invalid password.")
+        raise InvalidPasswordException("Invalid password.")
 
     if superuser:
         return User.objects.create_superuser(username, email, password, **kwargs)
@@ -51,7 +55,7 @@ def get_user(user: User, user_pk: int) -> User:
         raise NoPermissionException("No permission.")
     r_user = User.objects.filter(pk=user_pk).first()
     if not r_user:
-        raise UserNotFoundException()
+        raise UserNotFoundException("User not found.")
     return r_user
 
 
@@ -70,9 +74,17 @@ def update_user(
     with transaction.atomic():
         target_user = User.objects.filter(pk=user_pk).select_for_update().first()
         if not target_user:
-            raise UserNotFoundException()
+            raise UserNotFoundException("User not found.")
 
         target_user.email = email
+        if user == target_user:
+            if user.is_superuser != superuser:
+                raise InvalidOperationRequestException(
+                    "Can't upgrade/downgrade yourself."
+                )
+            if "is_active" in kwargs and kwargs["is_active"] != target_user.is_active:
+                raise InvalidOperationRequestException("Can't deactivate yourself.")
+
         target_user.is_superuser = superuser
         for key, value in kwargs.items():
             setattr(target_user, key, value)
@@ -85,6 +97,8 @@ def delete_user(user: User, user_pk: int):
     """Delete user"""
     if not user.is_superuser:
         raise NoPermissionException("No permission.")
+    if user.id == user_pk:
+        raise InvalidOperationRequestException("Can't delete yourself.")
 
     count, _ = User.objects.filter(pk=user_pk).all().delete()
     if not count:
